+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-08-17
+
 ### Added
 
 - Add `focusTrap` helper
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Everything!
 
-[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.0...v0.1.1